FIX: fix file paths
