+++ pyfibre/cli/tests/test_click_app.py
@@ -4,7 +4,7 @@ from unittest import mock, TestCase
 
 from click.testing import CliRunner
 
-import pyfibre.cli.pyfibre_cli
+import pyfibre.cli.app
 from pyfibre.version import __version__
 
 
@@ -13,7 +13,7 @@ def mock_run_pyfibre():
 
 
 def mock_run_constructor(*args, **kwargs):
-    mock_pyfibre_app = mock.Mock(spec=pyfibre.cli.pyfibre_cli)
+    mock_pyfibre_app = mock.Mock(spec=pyfibre.cli.app)
     mock_pyfibre_app.run = lambda: None
 
 
@@ -21,30 +21,30 @@ class TestClickRun(TestCase):
 
     def test_click_cli_version(self):
         clirunner = CliRunner()
-        clirunner.invoke(pyfibre.cli.pyfibre_cli.pyfibre,
+        clirunner.invoke(pyfibre.cli.app.pyfibre,
                          args="--version")
 
-    def test_click_gui_main(self):
+    def test_click_cli_main(self):
 
-        with mock.patch('pyfibre.cli.pyfibre_cli') as mock_run:
+        with mock.patch('pyfibre.cli.app') as mock_run:
             mock_run.side_effect = mock_run_constructor
 
-            pyfibre.cli.pyfibre_cli.pyfibre()
+            pyfibre.cli.app.pyfibre()
 
             self.assertTrue(mock_run.pyfibre.called)
 
     def test_run(self):
         with mock.patch(
-                'pyfibre.cli.app'
+                'pyfibre.cli.pyfibre_cli'
                 '.PyFibreApplication._run_pyfibre') as mock_pyfibre:
             mock_pyfibre.side_effect = mock_run_pyfibre
-            pyfibre.cli.pyfibre_cli.run(
+            pyfibre.cli.app.run(
                 file_path='', key='', sigma=0.5, alpha=0.5,
                 log_name='pyfibre', database_name='', debug=False,
                 profile=False, ow_metric=False, ow_segment=False,
                 ow_network=False, save_figures=False, test=False
             )
-            self.log = pyfibre.cli.pyfibre_cli.logging.getLogger(__name__)
+            self.log = pyfibre.cli.app.logging.getLogger(__name__)
             # This test seems to be broken at the moment
             # self.assertEqual(10, self.log.getEffectiveLevel())
             if os.path.exists('pyfibre.log'):
@@ -52,10 +52,10 @@ class TestClickRun(TestCase):
 
     def test_run_with_profile(self):
         with mock.patch(
-                'pyfibre.cli.app'
+                'pyfibre.cli.pyfibre_cli'
                 '.PyFibreApplication._run_pyfibre') as mock_pyfibre:
             mock_pyfibre.side_effect = mock_run_pyfibre
-            pyfibre.cli.pyfibre_cli.run(
+            pyfibre.cli.app.run(
                 file_path='', key='', sigma=0.5, alpha=0.5,
                 log_name='pyfibre', database_name='', debug=False,
                 profile=True, ow_metric=False, ow_segment=False,
